Remove ntp and clock related logic / multiple measurements before publishing (#3)
* refactor: replace measurement interval with awake duration and remove clock /ntp dependency

* refactor: adjust sensor data types and improve moisture/battery reading logic

* chore: remove unused sntpc dependency and update documentation

* refactor: skip battery voltage reading when USB is connected

+++ src/domain.rs
@@ -42,8 +42,8 @@ pub enum Sensor {
     WaterLevel(WaterLevel),
     AirTemperature(i16),
     AirHumidity(u16),
-    SoilMoisture(u16),
-    BatteryVoltage(u32),
+    SoilMoisture(u8),
+    BatteryVoltage(u16),
     SoilMoistureRaw(u16),
 }
